requirements_for_docker.txt
# --------- pytorch --------- #
pytorch-lightning #==1.9.2
torchmetrics #>=0.11.4

# --------- ---------------#
pandas #==2.0.2
matplotlib #==3.7.1
scikit-learn #==1.2.2
scikit-image #==0.21.0
tqdm #==4.65.0
Pillow
opencv-python
albumentations
einops

# --------- hydra --------- #
hydra-core #==1.3.2
hydra-colorlog #==1.2.0
hydra-optuna-sweeper #==1.2.0

# --------- loggers --------- #
wandb
# neptune-client
# mlflow
# comet-ml

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)
kaggle          # to download dataset

# --------resolve conflic ------------#
markdown-it-py==2.0.0
mdit-py-plugins
jupytext
tornado
cmake
lit
